fix: replace deprecated lsp plugin
jose-elias-alvarez/null-ls.nvim has been archived on Aug 12, 2023 and is no longer maintained.
The use of this plugin causes a bunch of warnings in neovim because of some functions being
deprecated that are used in the plugin.

Switch to nvimtools/none-ls.nvim which is the successor of the previous plugin.
It uses the same API, so it shouldn't affect the functionality in which null-ls.nvim
was previously utilized.